requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.10
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
idna==3.3
phonenumbers==8.12.40
phonenumberslite==8.12.40
pkg-resources==0.0.0
PyJWT==1.7.1
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
twilio==7.4.0
typing-extensions==4.0.1
urllib3==1.26.7
